Created a NewMessage component to avoid rendering the entire MessagesScreen
